Move `diff_between_datetimes` into its own file and include appropriate license
