+++ src/builtins/time.rs
@@ -16,6 +16,8 @@ use chrono::{
 };
 use chrono_tz::Tz;
 
+mod diff;
+
 pub fn register(m: &mut HashMap<&'static str, builtins::BuiltinFcn>) {
     m.insert("time.add_date", (add_date, 4));
     m.insert("time.clock", (clock, 1));
@@ -92,8 +94,6 @@ fn date(span: &Span, params: &[Ref<Expr>], args: &[Value], _strict: bool) -> Res
     .into())
 }
 
-// Adapted from the official Go implementation:
-// https://github.com/open-policy-agent/opa/blob/eb17a716b97720a27c6569395ba7c4b7409aae87/topdown/time.go#L179-L243
 fn diff(span: &Span, params: &[Ref<Expr>], args: &[Value], _strict: bool) -> Result<Value> {
     let name = "time.diff";
     ensure_args_count(span, name, params, args, 2)?;
@@ -101,49 +101,7 @@ fn diff(span: &Span, params: &[Ref<Expr>], args: &[Value], _strict: bool) -> Res
     let (datetime1, _) = parse_epoch(name, &params[0], &args[0])?;
     let (datetime2, _) = parse_epoch(name, &params[1], &args[1])?;
 
-    // Make sure both datetimes in the same timezone
-    let datetime2 = datetime2.with_timezone(&datetime1.timezone());
-
-    // Make sure `datetime1` is always the smallest one
-    let (datetime1, datetime2) = if datetime1 > datetime2 {
-        (datetime2, datetime1)
-    } else {
-        (datetime1, datetime2)
-    };
-
-    let mut year = datetime2.year() - datetime1.year();
-    let mut month = datetime2.month() as i32 - datetime1.month() as i32;
-    let mut day = datetime2.day() as i32 - datetime1.day() as i32;
-    let mut hour = datetime2.hour() as i32 - datetime1.hour() as i32;
-    let mut min = datetime2.minute() as i32 - datetime1.minute() as i32;
-    let mut sec = datetime2.second() as i32 - datetime1.second() as i32;
-
-    // Normalize negative values
-    if sec < 0 {
-        sec += 60;
-        min -= 1;
-    }
-    if min < 0 {
-        min += 60;
-        hour -= 1;
-    }
-    if hour < 0 {
-        hour += 24;
-        day -= 1;
-    }
-    if day < 0 {
-        // Days in month:
-        let t = Utc
-            .with_ymd_and_hms(datetime1.year(), datetime1.month(), 32, 0, 0, 0)
-            .single()
-            .ok_or(anyhow!("Could not convert `ns1` to datetime"))?;
-        day += 32 - t.day() as i32;
-        month -= 1;
-    }
-    if month < 0 {
-        month += 12;
-        year -= 1;
-    }
+    let (year, month, day, hour, min, sec) = diff::diff_between_datetimes(datetime1, datetime2)?;
 
     Ok(Vec::from([
         (year as i64).into(),
